rename projects component folder to project
