+++ frontend/components/project/ProjectDetails.vue
