+++ frontend/components/project/ResourceDirectory.vue
