+++ frontend/components/project/resources/Empty.vue
