+++ frontend/components/project/resources/ResourceFile.vue
