+++ frontend/components/project/resources/ResourceFolder.vue
