+++ frontend/pages/projects.vue
@@ -17,7 +17,7 @@
 </template>
 <script>
 import AddResourceModal from './../components/modals/AddResourceModal'
-import IndividualProjectView from './../components/projects/IndividualProjectView'
+import IndividualProjectView from './../components/project/IndividualProjectView'
 
 export default {
   components: { IndividualProjectView, AddResourceModal },
